Updated en_us.json
Renamed 'bg_looses_his_mind' back to 'C418 - 13'. This was renamed early on since I thought it'd be funny to change the sound file for the 13 disc to the entire bg rant, however the sound file was removed later in development. I just forgot to change the name back in the lang file.