Re-added Emerald Ore Retexture & Emmissive Texture
I made this retexture back when I first started development back in July. The idea with this texture is that it's an ancient, shattered jar o' enchanting that's been scattered throughout the stone. Don't remember why I removed it but I'd like to re-add it in this release since I think it fits with the emerald re-textures well.